+++ lib/app_widgets.dart
@@ -1,2 +1,3 @@
 export 'package:pactus_gui_widgetbook/src/features/widgets/chip_text_box/chip_text_box_widget.dart';
 export 'package:pactus_gui_widgetbook/src/features/widgets/buttons/adaptive_primary_button/adaptive_primary_button.dart';
+export 'package:pactus_gui_widgetbook/src/features/widgets/buttons/adaptive_secondary_button/adaptive_secondary_button.dart';